+++ next.config.js
@@ -8,6 +8,7 @@ module.exports = {
         SERVICE_ENCRYPTION_IV: process.env.SERVICE_ENCRYPTION_IV,
         SERVICE_ENCRYPTION_KEY: process.env.SERVICE_ENCRYPTION_KEY,
         SERVICE_DECRYPTION_KEY: process.env.SERVICE_DECRYPTION_KEY,
+        JWT_SECRET: process.env.JWT_SECRET,
     },
     async redirects() {
         return [
